gesture-recognition-system/requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.9
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.35.1
pyserial==3.5
